+++ doran/src/components/MainPage/PostCard.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+
+export default function PostCard() {
+  return <>
+  
+  </>;
+}
+
