--- doran/src/components/home/PostCard.tsx
@@ -1,26 +1,0 @@
-import React from "react";
-
-export default function PostCard() {
-  return <>
-  
-  </>;
-}
-
-/**
- * 📱 컴포넌트 특징:
- *
- * ✨ 현대적 디자인 요소
- * - 깔끔한 카드 레이아웃
- * - 부드러운 그라데이션과 그림자
- * - 일관된 간격과 타이포그래피
- *
- * 🎯 인터랙티브 요소
- * - 호버 애니메이션
- * - 버튼 상태 변화
- * - 접근성을 고려한 ARIA 라벨
- *
- * 📐 반응형 디자인
- * - 모바일 최적화
- * - 적절한 터치 타겟 크기
- * - 읽기 편한 타이포그래피
- */
